refactor(folder): 폴더 조회와 액션 훅 분리
