+++ src/features/clip/ui/ClipEntryPage.tsx
@@ -4,7 +4,7 @@ import type { ReactNode } from "react";
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { ClipListSkeleton } from "@/features/clip/ui/ClipListSkeleton";
-import { useFolders } from "@/features/folder/hooks/useFolders";
+import { useFoldersQuery } from "@/features/folder/hooks/useFoldersQuery";
 
 // 기본 컬렉션 경로를 첫 폴더 경로로 연결하고 폴더가 없으면 대체 화면을 표시합니다.
 interface ClipEntryPageProps {
@@ -14,7 +14,7 @@ interface ClipEntryPageProps {
 
 export function ClipEntryPage({ fallback, section }: ClipEntryPageProps) {
   const router = useRouter();
-  const { folders, isLoading } = useFolders();
+  const { folders, isLoading } = useFoldersQuery();
 
   useEffect(() => {
     if (!folders.length) {
