+++ src/features/folder/service/folderQueryCache.ts
@@ -1,0 +1,4 @@
+export const FOLDER_QUERY_KEY = "folders";
+
+export const getFolderQueryKey = (userId: string | null) =>
+  [FOLDER_QUERY_KEY, userId] as const;
